// artzt-patient-app/src/app/models/Formular.model.ts
export class Formular {

  formId: string;
  date: string;
  mood: string;
  moodText: string;
  activityLevel: string;
  morningLow: string;
  sleepQuality: string;
  sleepDisorder: string;
  appetit: string;
  otherBodySymptoms: string;
  thoughtsOfSuicide: string;
  drugs: string[] = [];
  dosesOfMedication: string[] = [];
}